refactor: organise frontend, add utils and centralise constants
