+++ frontend/src/Data/VideoGenre.js
@@ -1,0 +1,22 @@
+export const genres = [
+  { label: "Gaming", key: "gaming" },
+  { label: "Education", key: "education" },
+  { label: "Music", key: "music" },
+  { label: "Comedy", key: "comedy" },
+  { label: "Vlogs", key: "vlogs" },
+  { label: "Tech", key: "tech" },
+  { label: "Fitness", key: "fitness" },
+  { label: "Food", key: "food" },
+  { label: "Travel", key: "travel" },
+  { label: "Lifestyle", key: "lifestyle" },
+  { label: "DIY", key: "diy" },
+  { label: "Beauty", key: "beauty" },
+  { label: "Movies", key: "movies" },
+  { label: "Movie Reviews", key: "movie_reviews" },
+  { label: "ASMR", key: "asmr" },
+  { label: "Podcasts", key: "podcasts" },
+  { label: "Finance", key: "finance" },
+  { label: "Science", key: "science" },
+  { label: "Sports", key: "sports" },
+  { label: "Motivation", key: "motivation" },
+];
