+++ frontend/src/Data/VideoType.js
@@ -1,0 +1,4 @@
+export const videoType = [
+  { key: "video", label: "Video" },
+  { key: "short", label: "Short" },
+];
