+++ frontend/src/config/api.js
@@ -1,0 +1,2 @@
+export const API_BASE_URL =
+  `http://${import.meta.env.VITE_IP_ADD}:8000`;
